Reorganize to deduplicate code and avoid weird imports
