+++ explain.py
@@ -1,4 +1,4 @@
-import json
+from .explanations import il_explanations as explanations
 import traceback
 
 from binaryninja import (
@@ -22,15 +22,6 @@ no_paren = [
     LowLevelILOperation.LLIL_POP,
     LowLevelILOperation.LLIL_FLAG,
 ]
-
-# Using user_plugin_path doesn't work with plugins that have been installed from the repository manager,
-# since it points to .binaryninja/plugins instead of .binaryninja/repositories
-# path = user_plugin_path + '/binja_explain_instruction/explanations_en.json'
-import os
-
-path = os.path.dirname(os.path.realpath(__file__)) + "/explanations_en.json"
-with open(path, "r") as explanation_file:
-    explanations = json.load(explanation_file)
 
 
 def preprocess_LLIL_CONST(_bv, llil_instruction):
