+++ explainers/__init__.py
@@ -1,0 +1,8 @@
+# Import as lowercase here so we can use string formatting to import them later
+from .x86 import X86Explainer as x86_explainer
+from .mips import MipsExplainer as mips_explainer
+from .aarch64 import AArch64Explainer as aarch64_explainer
+from .ual import UALExplainer as ual_explainer
+from .asm6502 import Asm6502Explainer as asm6502_explainer
+from .msp430 import MSP430Explainer as msp430_explainer
+from .powerpc import PowerPCExplainer as powerpc_explainer
