+++ explainers/mips.py
@@ -1,0 +1,6 @@
+from .generic import GenericExplainer
+from ..instruction_docs import mips_instructions
+
+
+class MipsExplainer(GenericExplainer):
+    instruction_docs = mips_instructions
