+++ explainers/msp430.py
@@ -1,0 +1,6 @@
+from .generic import GenericExplainer
+from ..instruction_docs import msp430_instructions
+
+
+class MSP430Explainer(GenericExplainer):
+    instruction_docs = msp430_instructions
