+++ explainers/powerpc.py
@@ -1,0 +1,6 @@
+from .generic import GenericExplainer
+from ..instruction_docs import powerpc_instructions
+
+
+class PowerPCExplainer(GenericExplainer):
+    instruction_docs = powerpc_instructions
