+++ gui.py
@@ -1,6 +1,5 @@
 import traceback
 import typing
-from importlib import import_module
 
 from PySide6.QtCore import Qt
 from PySide6.QtGui import QFontDatabase, QFont
@@ -11,18 +10,19 @@ from PySide6.QtWidgets import QVBoxLayout, QLabel
 from binaryninja import BinaryView, Architecture, log_error
 from binaryninjaui import SidebarWidget, UIActionHandler
 
+from .util import rec_replace
 from .explain import explain_llil, fold_multi_il
-from .util import *
 from .util import (
     get_function_at,
     find_llil,
     find_lifted_il,
     inst_in_func,
-    dereference_symbols,
 )
 
-arch_specific_explanations = {}
-doc_urls = {}
+from . import explainers
+from .explainers.generic import UnavailableExplainer
+
+arch_explainer_constructors = {}
 _remappings = {
     "arm": "ual",
     "thumb2": "ual",
@@ -41,14 +41,9 @@ for arch in Architecture:
         "powerpc",
     ):
         if supported in arch.name:
-            doc_urls[arch.name] = import_module(
-                f".{_remappings.get(supported, supported)}",
-                package="binja_explain_instruction",
-            ).get_doc_url
-            arch_specific_explanations[arch.name] = import_module(
-                f".{_remappings.get(supported, supported)}.explain",
-                package="binja_explain_instruction",
-            ).arch_explain_instruction
+            arch_explainer_constructors[arch.name] = getattr(
+                explainers, f"{_remappings.get(supported, supported)}_explainer"
+            )
 
 
 def make_hline():
@@ -56,19 +51,6 @@ def make_hline():
     out.setFrameShape(QFrame.HLine)
     out.setFrameShadow(QFrame.Sunken)
     return out
-
-
-def no_documentation_available(*_args):
-    return [
-        (
-            "No documentation available",
-            "https://github.com/ehennenfent/binja_explain_instruction/blob/master/CONTRIBUTING.md",
-        )
-    ]
-
-
-def no_arch_specific_explanations_available(*_args):
-    return False, ["Architecture-specific explanations unavailable"]
 
 
 class ExplanationWindow(SidebarWidget):
@@ -81,13 +63,10 @@ class ExplanationWindow(SidebarWidget):
 
         self.configured_arch = None
         self._bv = None
+        self.arch_explainer = None
 
+        # Configures configured_arch, _bv, and arch_explainer
         self.bv = bv
-
-        self.architecture_specific_explanation_function = (
-            no_arch_specific_explanations_available
-        )
-        self.get_doc_url = no_documentation_available
 
         self._layout = QVBoxLayout(self)
 
@@ -203,9 +182,7 @@ class ExplanationWindow(SidebarWidget):
         (
             should_supersede_llil,
             explanation_list,
-        ) = self.architecture_specific_explanation_function(
-            self.bv, instruction, lifted_il_list
-        )
+        ) = self.arch_explainer.explain_instruction(instruction, lifted_il_list)
 
         # Display the raw instruction
         try:
@@ -229,7 +206,7 @@ class ExplanationWindow(SidebarWidget):
             # We append the line number if we're displaying a conditional.
             self.description = [explain_llil(self.bv, llil) for llil in parse_il]
 
-        # Display the  LLIL, dereferencing anything that looks like a hex number into a symbol if possible
+        # Display the LLIL, dereferencing anything that looks like a hex number into a symbol if possible
         self.llil = (
             llil_list  # [dereference_symbols(self.bv, llil) for llil in llil_list]
         )
@@ -241,12 +218,9 @@ class ExplanationWindow(SidebarWidget):
 
     def configure_for_arch(self, arch: Architecture):
         self.configured_arch = arch
-        self.architecture_specific_explanation_function = (
-            arch_specific_explanations.get(
-                arch.name, no_arch_specific_explanations_available
-            )
-        )
-        self.get_doc_url = doc_urls.get(arch.name, no_documentation_available)
+        self.arch_explainer = arch_explainer_constructors.get(
+            arch.name, UnavailableExplainer
+        )(self.bv)
 
     def notifyOffsetChanged(self, offset):
         self.explain_instruction(offset)
@@ -260,3 +234,98 @@ class ExplanationWindow(SidebarWidget):
 
     def contextMenuEvent(self, event):
         self.m_contextMenuManager.show(self.m_menu, self.actionHandler)
+
+
+def parse_instruction(context, instr):
+    """Helps the GUI go from lists of instruction data to a cleanly formatted string"""
+    if instr is not None:
+        docs = context.arch_explainer.get_doc_url(instr.split(" "))
+        instruction = context.escape(instr.replace("    ", " "))
+        shortForm = context.newline.join(
+            '<a href="{href}">{form}</a>'.format(
+                href=url, form=context.escape(short_form)
+            )
+            for short_form, url in docs
+        )
+        return instruction, shortForm
+    else:
+        return "None", "None"
+
+
+def parse_description(context, desc_list):
+    return context.newline.join(
+        context.escape(new_description) for new_description in desc_list
+    )
+
+
+def parse_llil(context, llil_list):
+    """Helps the GUI go from lists of instruction data to a cleanly formatted string"""
+    newText = ""
+    for llil in llil_list:
+        if llil is not None:
+            tokens = llil.deref_tokens if hasattr(llil, "deref_tokens") else llil.tokens
+            newText += "{}: ".format(llil.instr_index)
+            newText += "".join(context.escape(str(token)) for token in tokens)
+        else:
+            newText += "None"
+        newText += context.newline
+    if len(llil_list) > 0:
+        return newText.strip(context.newline)
+    else:
+        return "None"
+
+
+def parse_mlil(context, mlil_list):
+    """Helps the GUI go from lists of instruction data to a cleanly formatted string"""
+    newText = ""
+    for mlil in mlil_list:
+        if mlil is not None:
+            tokens = mlil.deref_tokens if hasattr(mlil, "deref_tokens") else mlil.tokens
+            newText += "{}: ".format(mlil.instr_index)
+            newText += "".join(context.escape(str(token)) for token in tokens)
+        else:
+            newText += "None"
+        newText += context.newline
+    if len(mlil_list) > 0:
+        return newText.strip(context.newline)
+    else:
+        return "None"
+
+
+def parse_state(context, state_list):
+    if state_list is not None:
+        return context.newline.join(context.escape(state) for state in state_list)
+    else:
+        return "None"
+
+
+def parse_flags(context, tuple_list_list):
+    """Helps the GUI go from lists of instruction data to a cleanly formatted string"""
+    out = ""
+    for f_read, f_written, lifted in tuple_list_list:
+        if len(f_read) > 0:
+            out += (
+                (
+                    "(Lifted IL: {}) ".format(lifted.instr_index)
+                    if len(tuple_list_list) > 1
+                    else ""
+                )
+                + "Reads: "
+                + ", ".join(f_read)
+                + context.newline
+            )
+        if len(f_written) > 0:
+            out += (
+                (
+                    "(Lifted IL: {}) ".format(lifted.instr_index)
+                    if len(tuple_list_list) > 1
+                    else ""
+                )
+                + "Writes: "
+                + ", ".join(f_written)
+                + context.newline
+            )
+        out += context.newline
+    out = rec_replace(out.strip(context.newline), context.newline * 2, context.newline)
+    out = "None" if len(out) == 0 else out
+    return out
