+++ instruction_docs/__init__.py
@@ -1,0 +1,7 @@
+from .aarch64 import instrs as aarch64_instructions
+from .asm6502 import instrs as asm6502_instructions
+from .mips import instrs as mips_instructions
+from .msp430 import instrs as msp430_instructions
+from .powerpc import instrs as powerpc_instructions
+from .ual import instrs as ual_instructions
+from .x86 import instrs as x86_instructions
