+++ instruction_docs/powerpc.py
@@ -2492,26 +2492,3 @@ instrs = {
         }
     ],
 }
-
-
-def find_proper_name(instruction):
-    out = str(instruction).strip().lower()
-    return out
-
-
-def get_doc_url(i):
-    """Takes in the instruction tokens and returns [(short form, doc url)]"""
-    names = map(find_proper_name, i)  # handles instruction prefixes
-    output = []
-    for name in names:
-        if name in instrs.keys():
-            inst_data = instrs[name][0]
-            output.append((inst_data["short"], inst_data["link"]))
-    if len(output) == 0:
-        return [
-            (
-                "No documentation for that instruction!",
-                "https://github.com/ehennenfent/binja_explain_instruction/blob/master/CONTRIBUTING.md",
-            )
-        ]
-    return output
