--- msp430/explanations_en.json
@@ -1,1 +1,0 @@
-{}
