+++ util.py
@@ -52,69 +52,6 @@ def dereference_symbols(bv, il_instruction):
     return il_instruction
 
 
-def parse_instruction(context, instr):
-    """Helps the GUI go from lists of instruction data to a cleanly formatted string"""
-    if instr is not None:
-        docs = context.get_doc_url(instr.split(" "))
-        instruction = context.escape(instr.replace("    ", " "))
-        shortForm = context.newline.join(
-            '<a href="{href}">{form}</a>'.format(
-                href=url, form=context.escape(short_form)
-            )
-            for short_form, url in docs
-        )
-        return instruction, shortForm
-    else:
-        return "None", "None"
-
-
-def parse_description(context, desc_list):
-    return context.newline.join(
-        context.escape(new_description) for new_description in desc_list
-    )
-
-
-def parse_llil(context, llil_list):
-    """Helps the GUI go from lists of instruction data to a cleanly formatted string"""
-    newText = ""
-    for llil in llil_list:
-        if llil is not None:
-            tokens = llil.deref_tokens if hasattr(llil, "deref_tokens") else llil.tokens
-            newText += "{}: ".format(llil.instr_index)
-            newText += "".join(context.escape(str(token)) for token in tokens)
-        else:
-            newText += "None"
-        newText += context.newline
-    if len(llil_list) > 0:
-        return newText.strip(context.newline)
-    else:
-        return "None"
-
-
-def parse_mlil(context, mlil_list):
-    """Helps the GUI go from lists of instruction data to a cleanly formatted string"""
-    newText = ""
-    for mlil in mlil_list:
-        if mlil is not None:
-            tokens = mlil.deref_tokens if hasattr(mlil, "deref_tokens") else mlil.tokens
-            newText += "{}: ".format(mlil.instr_index)
-            newText += "".join(context.escape(str(token)) for token in tokens)
-        else:
-            newText += "None"
-        newText += context.newline
-    if len(mlil_list) > 0:
-        return newText.strip(context.newline)
-    else:
-        return "None"
-
-
-def parse_state(context, state_list):
-    if state_list is not None:
-        return context.newline.join(context.escape(state) for state in state_list)
-    else:
-        return "None"
-
-
 def rec_replace(in_str, old, new):
     """Recursively replace a string in a string"""
     if old == new:
@@ -124,33 +61,19 @@ def rec_replace(in_str, old, new):
     return rec_replace(in_str.replace(old, new), old, new)
 
 
-def parse_flags(context, tuple_list_list):
-    """Helps the GUI go from lists of instruction data to a cleanly formatted string"""
-    out = ""
-    for f_read, f_written, lifted in tuple_list_list:
-        if len(f_read) > 0:
-            out += (
-                (
-                    "(Lifted IL: {}) ".format(lifted.instr_index)
-                    if len(tuple_list_list) > 1
-                    else ""
-                )
-                + "Reads: "
-                + ", ".join(f_read)
-                + context.newline
-            )
-        if len(f_written) > 0:
-            out += (
-                (
-                    "(Lifted IL: {}) ".format(lifted.instr_index)
-                    if len(tuple_list_list) > 1
-                    else ""
-                )
-                + "Writes: "
-                + ", ".join(f_written)
-                + context.newline
-            )
-        out += context.newline
-    out = rec_replace(out.strip(context.newline), context.newline * 2, context.newline)
-    out = "None" if len(out) == 0 else out
-    return out
+class AttrDict(dict):
+    """Borrowed from https://stackoverflow.com/a/14620633. Lets us use the . notation
+    in format strings."""
+
+    def __init__(self, *args, **kwargs):
+        super(AttrDict, self).__init__(*args, **kwargs)
+        self.__dict__ = self
+
+
+def parse_instruction(instruction):
+    """Removes whitespace and commas from the instruction tokens"""
+    tokens = filter(
+        lambda x: len(x) > 0,
+        [str(token).strip().replace(",", "") for token in str(instruction).split(" ")],
+    )
+    return list(tokens)
